Add online connectivity dependency to systemd service
If networking hasn't connected before server startup, the server won't be
able to assign itself an IP (as system still doesn't have one), and will
throw an exception. Systemd will then try to respawn the process several
times and then quit trying (especially if 'RestartSec' isn't set here),
all ending in a non-started server after a power loss (Raspberry is
connected to the Internet via Wi-Fi - takes a bit for the router to boot
up). This change (along with the 'RestartSec' one) should mitigate
this issue.